requirements.txt
# Core
torch>=1.7.0
torchvision>=0.8.0
numpy
opencv-python-headless

# SR Models
basicsr>=1.4.2
facexlib>=0.2.5
gfpgan>=1.3.5
realesrgan>=0.3.0

# Server
fastapi
uvicorn
pydantic
requests
python-multipart

# OSS & Utils
oss2
pyyaml
